app/requirements.txt
amqp==5.2.0
asgiref==3.6.0
beautifulsoup4==4.11.2
billiard==4.2.0
celery==5.3.4
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==3.1.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.2
defusedxml==0.7.1
Django==4.1.7
django-cors-headers==3.14.0
django-js-asset==2.0.0
django-mptt==0.14.0
django-taggit==3.1.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-recursive==0.1.2
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
gunicorn==20.1.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
kombu==5.3.3
MarkupSafe==2.1.2
oauthlib==3.2.2
pillow==10.2.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
setuptools==68.2.2
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
soupsieve==2.4
sqlparse==0.4.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
vine==5.1.0
wcwidth==0.2.9
wheel==0.41.2
